+++ lib/phantom/resource_template.ex
@@ -57,11 +57,55 @@ defmodule Phantom.ResourceTemplate do
   Build a resource_template spec
 
   The `Phantom.Router.resource/3` macro will build these specs.
+
+  - `:name` - The name of the resource template.
+  - `:uri` - The URI template of the resource in the style of `Plug.Router`, including the scheme.
+    For example, you can define a path like `"myapp:///some/path/:project_id/:id` which
+    will be parsed to include path params `%{"project_id" => _, "id" => _}`. The scheme can be
+    `"https"`, `"git"`, `"file"`, or custom, eg `"myapp"`.
+  - `:description` - The description of the resource and when to use it.
+  - `:handler` - The module to call.
+  - `:function` - The function to call on the handler module.
+  - `:completion_function` - The function to call on the handler module that will provide possible completion results.
+  - `:mime_type` - the MIME type of the results.
+  - `:router` - The Router module that will capture the URIs and route resources by URI to functions.
+    This is constructed by the `Phantom.Router.resource/3` macro automatically as
+    `MyApp.MyMCPRouter.ResourceRouter.{Scheme}`. The module does not need to exist at the time of
+    building it-- it will be generated when added by `Phantom.Cache.add_resource_template/2`
+    or by the `Phantom.Router.resource/3` macro.
+
   """
   def build(attrs) do
-    attrs = Map.new(attrs)
-    uri_template = "#{attrs.scheme}://#{to_uri_6570(attrs.path)}"
-    struct!(__MODULE__, Map.put(attrs, :uri_template, uri_template))
+    uri =
+      case attrs[:uri] do
+        %URI{} = uri -> {:ok, uri}
+        uri when is_binary(uri) -> URI.new(uri)
+      end
+
+    uri =
+      case uri do
+        {:ok, %{path: path, scheme: scheme} = uri}
+        when is_binary(path) and is_binary(scheme) ->
+          uri
+
+        {:ok, uri} ->
+          raise "Provided an invalid URI.\nResource URIs must contain a path and a scheme.\nProvided: #{URI.to_string(uri)}"
+
+        {:error, invalid} ->
+          raise "Provided an invalid URI.\nProvided: #{inspect(attrs[:uri])}\nError at: #{inspect(invalid)}"
+      end
+
+    struct!(
+      __MODULE__,
+      attrs
+      |> Map.new()
+      |> Map.merge(%{
+        name: attrs[:name] || to_string(attrs[:function]),
+        scheme: attrs[:scheme] || uri.scheme,
+        path: attrs[:path] || uri.path,
+        uri_template: "#{uri.scheme}://#{to_uri_6570(uri.path)}"
+      })
+    )
   end
 
   @spec to_json(t()) :: json()